api/v1/accounts/serializers.py
from django.contrib.auth.models import User, Group
from rest_framework import serializers


class LoginSerializer(serializers.Serializer):
    username = serializers.CharField()
    password = serializers.CharField()