requirements.txt
asgiref==3.7.2
# backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
# jwt==1.3.1
psycopg2-binary==2.9.6
pycparser==2.21
pytz==2023.3
requests==2.31.0
# rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
